DS-224 Fix problems with searching for "zijkanaal h 2" that is in the search tests
If dashes are replaced with pattern_replace that is interpreted differently for Zijkanaal H-weg 2
So this should be done instead with the mapping type char_filter

The quote character should also not be included in ther mapping replace. We can search
directly for the  quotei character in ES.

This will also fix DS-226 about UTF-8 Apostrophe because the asciifolding will take replace any
UTF characters by their ascii counterparts.