DS-448 - Make typeahead search for nummeraanduiding better.
First use the score for sorting the results, then the streetname an number.
In order to get consistent scoring we shoud do the search with :

search_type=dfs_query_then_fetch

See : https://www.elastic.co/guide/en/elasticsearch/reference/master/consistent-scoring.html

Then the search is also changed to do one of match or match_prefix.
This will put complete matches on top. For example Amstel 1

Al this is being made available as optional feature.
It can be turned on by using features=1

If it is OK we can make this the default, or remove the old version