Always include Weesp in search results.
Previously, including Weesp was depending on the `features=2`
request parameter. However, Weesp now just needs to be always included.

This has only been changed by disabling the feature flag.
The handling in the construction of the elasticsearch queries
is not touched, because this is quite complex.
Better not introduce new bugs :-)

This could be done when search is being migrated to Azure.