Add api key server middleware + config
The APIs in this repo will get an API Key check.
Initially the API Key is not mandatory.

The API key check is done by a Django middleware (datadiensten-apikeysclient)
and the configuration that is needed is added to the Django settings.

The public signing keys will be added to the container during deployment
from the Openstack ansible playbook.